delete mimic_samp if we userns

+++ brave.sh
@@ -5,7 +5,7 @@ mimic_stamp="$XDG_DATA_HOME/flatpak-brave-mimic-stamp"
 if [ ! -f "$mimic_stamp" ] && ! zypak-helper spawn-strategy-test; then
 	zenity --info --title='Brave Flatpak' --no-wrap --text="$(< /app/share/flatpak-brave/mimic_warning.txt)"
 	> "$mimic_stamp"
-elif [ -f "$mimic_stamp" ]; then
+elif [ -f "$mimic_stamp" ] && zypak-helper spawn-strategy-test; then
 	rm -f "$mimic_stamp"
 fi
